Created Training subcomponent
